challan

+++ UI/SAD/Order/RemoteAttach.aspx.cs
@@ -11,6 +11,7 @@ using System.Web;
 using System.Web.UI;
 using System.Web.UI.WebControls;
 using UI.ClassFiles;
+using GLOBAL_BLL;
 
 namespace UI.SAD.Order
 {
@@ -68,18 +69,37 @@ namespace UI.SAD.Order
                 string curentdate = DateTime.Now.ToString(".fffffff").ToString();
                 string dte = DateTime.Now.ToString(".fffffff").ToString();
 
-                string dfile = enrol + attachname + dte + Path.GetFileName(DUpload.PostedFile.FileName);
-                decimal length = dfile.Length;
-                path = dfile;
+                string Dfiles = DUpload.PostedFile.FileName;
+                string FileExtension = Dfiles.Substring(Dfiles.LastIndexOf('.') + 1).ToLower();
+                if (FileExtension == "jpeg" || FileExtension == "jpg" || FileExtension == "png")
+                {
 
-              
-                DUpload.PostedFile.SaveAs(Server.MapPath("~/SAD/Order/Data/OR/") + dfile);
-                FileUploadFTP(Server.MapPath("~/SAD/Order/Data/OR/"), dfile, "ftp://ftp.akij.net/TADAPictures/", "erp@akij.net", "erp123");
-                File.Delete(Server.MapPath("~/SAD/Order/Data/OR/") + dfile);
-                Int32 intPart = 1;
-                bll.getTADAAttachinsertion(dfile, length, path, enrol, unit, dteFromDate, AttachemtTypeid, jobstation, intPart);
-                ScriptManager.RegisterStartupScript(Page, typeof(Page), "StartupScript", "alert('Successfully Uploaded Bill docuement');", true);
 
+                    string dfile = enrol + attachname + dte + Path.GetFileName(DUpload.PostedFile.FileName);
+                    decimal length = dfile.Length;
+                    path = dfile;
+
+
+
+
+                    DocumentUpload_BLL objDocUp = new DocumentUpload_BLL();
+                    Stream strm = DUpload.PostedFile.InputStream;
+                    objDocUp.ImageCompress(strm, Server.MapPath("~/SAD/Order/Data/OR/") + dfile);
+
+
+                    //DUpload.PostedFile.SaveAs(Server.MapPath("~/SAD/Order/Data/OR/") + dfile);
+                    FileUploadFTP(Server.MapPath("~/SAD/Order/Data/OR/"), dfile, "ftp://ftp.akij.net/TADAPictures/", "erp@akij.net", "erp123");
+                    File.Delete(Server.MapPath("~/SAD/Order/Data/OR/") + dfile);
+                    Int32 intPart = 1;
+                    bll.getTADAAttachinsertion(dfile, length, path, enrol, unit, dteFromDate, AttachemtTypeid, jobstation, intPart);
+                    ScriptManager.RegisterStartupScript(Page, typeof(Page), "StartupScript", "alert('Successfully Uploaded Bill docuement');", true);
+                }
+
+                else
+                {
+                    ScriptManager.RegisterStartupScript(Page, typeof(Page), "StartupScript", "alert('Please  save as jpeg OR jpg or png format image ');", true);
+
+                }
 
 
             }
